// apps/web/src/lib/api/model/storeApprovalFlowRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Procurement
 * OpenAPI spec version: 0.0.1
 */
import type { StoreApprovalFlowRequestStepsItem } from './storeApprovalFlowRequestStepsItem';
import type { StoreApprovalFlowRequestSubjectType } from './storeApprovalFlowRequestSubjectType';

export interface StoreApprovalFlowRequest {
  /** @maxLength 64 */
  subject_type: StoreApprovalFlowRequestSubjectType;
  /** @maxLength 255 */
  name: string;
  /** @minimum 0 */
  min_amount: number;
  /**
     * @minimum 0
     * @nullable
     */
  max_amount?: number | null;
  is_active?: boolean;
  /** @minItems 1 */
  steps: StoreApprovalFlowRequestStepsItem[];
}