// apps/web/src/lib/api/model/storeApprovalFlowRequestSubjectType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Procurement
 * OpenAPI spec version: 0.0.1
 */

/**
 * @maxLength 64
 */
export type StoreApprovalFlowRequestSubjectType = typeof StoreApprovalFlowRequestSubjectType[keyof typeof StoreApprovalFlowRequestSubjectType];


export const StoreApprovalFlowRequestSubjectType = {
  evaluation: 'evaluation',
} as const;