// apps/web/src/lib/api/model/updateApprovalFlowRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Procurement
 * OpenAPI spec version: 0.0.1
 */
import type { UpdateApprovalFlowRequestStepsItem } from './updateApprovalFlowRequestStepsItem';

export interface UpdateApprovalFlowRequest {
  /** @maxLength 255 */
  name?: string;
  /** @minimum 0 */
  min_amount?: number;
  /**
     * @minimum 0
     * @nullable
     */
  max_amount?: number | null;
  is_active?: boolean;
  /** @minItems 1 */
  steps?: UpdateApprovalFlowRequestStepsItem[];
}